refactor: inject admin edit initial data (#378)
