+++ src/app/manage/posts/[id]/preview/page.tsx
@@ -1,5 +1,5 @@
-import { cookies } from "next/headers";
 import { notFound } from "next/navigation";
+import { toAdminPostCookieHeader } from "../admin-post-cookie";
 import { fetchAdminPost } from "@entities/post";
 import { ApiResponseError } from "@shared/api";
 import { renderMarkdown } from "@shared/lib/markdown";
@@ -8,18 +8,6 @@ import { PostPreview } from "@widgets/admin-post-preview";
 interface PostPreviewPageProps {
   // Next.js 15: params is a Promise. await is a no-op in Next.js 14.
   params: Promise<{ id: string }>;
-}
-
-async function toCookieHeader(): Promise<string | undefined> {
-  // await is a no-op in Next.js 14 but required in Next.js 15 where cookies() returns a Promise.
-  const cookieStore = await cookies();
-  const sessionCookie = cookieStore.get("sessionId");
-
-  if (!sessionCookie) {
-    return undefined;
-  }
-
-  return `${sessionCookie.name}=${encodeURIComponent(sessionCookie.value)}`;
 }
 
 export default async function PostPreviewPage({
@@ -32,7 +20,7 @@ export default async function PostPreviewPage({
     notFound();
   }
 
-  const cookieHeader = await toCookieHeader();
+  const cookieHeader = await toAdminPostCookieHeader();
 
   try {
     const post = await fetchAdminPost(id, cookieHeader);
