+++ src/features/post-editor/index.ts
@@ -1,7 +1,10 @@
 export { MarkdownEditor } from "./ui/markdown-editor";
 export { MarkdownPreview } from "./ui/markdown-preview";
 export { PostForm } from "./ui/post-form";
-export type { PostFormValues } from "./ui/post-form";
+export {
+  mapPostToFormValues,
+  type PostFormValues,
+} from "./lib/post-form-values";
 export {
   resolvePreviewContent,
   uploadPendingImages,
