+++ src/features/post-editor/ui/post-form.tsx
@@ -36,6 +36,10 @@ import {
   insertMarkdownImage,
   insertMarkdownImages,
 } from "../lib/markdown-commands";
+import {
+  mapPostToFormValues,
+  type PostFormValues,
+} from "../lib/post-form-values";
 import { attachScrollSync } from "../lib/scroll-sync";
 import type { EditorView } from "@codemirror/view";
 import { AssetPickerModal } from "@entities/asset";
@@ -65,19 +69,6 @@ type SubmitIntent =
   | "archive"
   | "restore-draft"
   | "unpublish";
-
-export interface PostFormValues {
-  title: string;
-  categoryId: number | null;
-  tags: string[];
-  status: PostDetail["status"];
-  visibility: PostDetail["visibility"];
-  commentStatus: "open" | "locked" | "disabled";
-  thumbnailUrl: string;
-  summary: string;
-  description: string;
-  contentMd: string;
-}
 
 interface PostFormProps {
   mode?: "create" | "edit";
@@ -159,21 +150,6 @@ function buildPayload(values: PostFormValues): CreatePostBody {
     tags: values.tags.length > 0 ? values.tags : undefined,
     summary: normalizeOptionalText(values.summary),
     description: normalizeOptionalText(values.description),
-  };
-}
-
-function mapPostToFormValues(post: PostDetail): PostFormValues {
-  return {
-    title: post.title,
-    categoryId: post.categoryId,
-    tags: post.tags.map((tag) => tag.name),
-    status: post.status,
-    visibility: post.visibility,
-    commentStatus: post.commentStatus ?? "open",
-    thumbnailUrl: post.thumbnailUrl ?? "",
-    summary: post.summary ?? "",
-    description: post.description ?? "",
-    contentMd: post.contentMd,
   };
 }
